Changes to topf calculation
I made some changes to the way topf is calculated to ensure that the
scrollbar always reaches the bottom at the correct spot. I also upped
the default acell value to 100, as I found 3 to be too slow, and not at
all what I'm used to with my default scrollbar. Also removed the height
set on the .scrollYou class.